[ci skip] Restructure folders (part 2)
