+++ Sources/Enchants/HeroPower.cpp
